hgmo: stream JSON responses (bug 1303904); r=glob
Previously, we called json.dumps() for some web commands. This buffers
the JSON response in memory then sends it. That buffering can take a
long time if the returned document is large.

Because of differences in how the JSON encoder is implemented, this
improves throughput of sending JSON from ~315KB/s to >7MB/s on my
i7-6700K.

Test output for the mozbuildinfo command has changed because we no
longer send trailing whitespace (which is what the default JSONEncoder
does in Python 2.7).

Because of clownshoes that will be fixed in a subsequent commit, this
drops the time to first byte on
json-automationrelevance/5dddbefdf759f09b1411f33fa0920835b919fc81
against the mozilla-aurora repo from ~119s to ~80s.

MozReview-Commit-ID: 2V4Y3iWfZRe

--HG--
extra : rebase_source : 0f52a8edc6a6be0d5367f9190445f2a1ba37a0e5