hgmo: refactor loop iterating over revs (bug 1303904); r=glob
We'll soon introduce a second consumer of this result. So we save the
revset result to a variable to avoid redundant execution. We also buffer
the result in a list because we know we'll be iterating twice and this
will be more efficient than iterating a revset result type.

MozReview-Commit-ID: IjAK6qMV8eA

--HG--
extra : rebase_source : cd6f77626ae6283629d645643624c038e8b8cded