pushlog: API to cache pushlog lookups (bug 1303904); r=glob
The pushlog API is not conducive to performance because various
methods result in a SQLite database handle being opened for every
lookup. When doing things like looking up pushlog entries for
every node in a list, the overhead can really add up.

This commit introduces a context manager for populating a cache
to facilitate rapid node -> push lookups.

This commit is somewhat hacky. There are a number of alternatives
I'd prefer, such as a bulk lookup API or automatic caching. A bulk
lookup API isn't usable in the case of the automationrelevance web
command because the function that calls the pushlog API is per
changeset, not per collection. And I'm not yet willing to implement
automatic caching because caching is hard and I don't want to
invite bugs.

A context manager, however, allows us to create caches with scoped
lifetimes thus limiting the caching (and bugs related to it) to
consumers who opt into the cache. Yes, it's hacky. But it gets the
job done.

To prove the cache works, the automationrelevance web command has
been updated to use it.

This drops time to first byte on
json-automationrelevance/5dddbefdf759f09b1411f33fa0920835b919fc81 HTTP
requests against the mozilla-aurora repo from ~80s to ~1.5s. However,
there is still a major source of clownshoes for that request...

MozReview-Commit-ID: 37bP49sHxKZ

--HG--
extra : rebase_source : 00e84cb58664f46383836a7cf8a43990a7952fa2
extra : source : f069a429dcae582956325dfc81bb9e5a16176219