hghooks: hook to trigger cache population (bug 1358239); r=glob
Mozilla has historically had problems with the tags cache causing
performance problems. Most of these problems were resolved a few years
ago by a rewritten tags cache implementation in Mercurial. However,
for that cache to work it needs to be populated. And, the cache isn't
populated until something attempts to resolve a symbol that could be
a tag.

This is normally not a problem. On hgweb, the tags cache will be
updated on most page views. However, it can be problematic on the
master server.

Nothing in the regular push flow appears to access tags data. This
means that the tags cache may not be written on the master server.
It so happens that we have processes running on the master server
that don't have write privileges to repos. So while they may trigger
tags cache population, the cache I/O fails and the cache is never
written. Assuming the cache is never written, over time these processes
have to compute more and more tags data and over time the amount of CPU
required balloons. Enough time passes and alerts start to fire because
processes that should be quick are spending dozens of seconds computing
tags data.

This commit solves that problem by implementing a pre transaction close
hook that accesses tags data, thus ensuring the tags cache is up to
date. Other processes should never have to compute tags data again.
Furthermore, `hg pull` will transfer tags cache data. So populating
the cache before replication ensures that hgweb machines also don't
have to populate the cache. Finally, the bundles generated for
clone bundles should have current tags caches, ensuring that people
who clone from them don't need to regenerate the data. So ensuring
the tags cache is populated on the master server is full of wins.

The new hook is globally installed on the master server so all repos
benefit from tags cache generation.

Unless we pre-populate the tags cache on all repos, the first push to
a repo after this is enabled may be slow.

All subsequent pushes may slow down because of this hook. The tags
cache population time is proportional to the number of files in a
repo and the number of heads being pushed. However, the common case
is 1 head per push and Mercurial should cache the manifest for that
head, thus assuring rapid tags cache generation. The tags cache
generation time is recorded in blackbox logs. So if the logs show
this hook causes too much of a perf drain, we can look into alternate
mechanisms for populating the tags cache.

MozReview-Commit-ID: 5q9TuqO4JaS

--HG--
extra : rebase_source : e9ebf73643d4f67a19f8193286c4164138a3772a