overlay: change content checking in non-contiguous mode (bug 1375966); r=smacleod
Before this change, non-contiguous mode compared content against the
last overlayed changeset. Since we're skipping changesets in
non-contiguous mode, this check would almost certainly fail. The
(current) goal in non-contiguous mode is to support backouts. The
skipped changeset will be a backout. The first parent of the first
changeset to be overlayed should be the backout. So, for non-contiguous
mode, we change the content check to be against the first parent.

This isn't a complete fix. If the first parent isn't the backout, then
the content check will fail. But this is presumed to be true most of
the time for the current Servo backout system. Perfect is the enemy of
good.

MozReview-Commit-ID: 79AHD6SYztA