mozautomation: support discovering and persisting Firefox releases (bug 1380176); r=glob
Many have wanted to easily answer questions like:

* Give me a list of all Firefox Nightly builds
* What is the first Nightly this revision appeared in
* Where are the build and test logs for this Nightly

Some of these questions are difficult to answer because we don't
have a central database aggregating metadata for all Firefox builds.
Without a low-latency database, we can't efficiently answer these
queries. That means we can't expose this data with low effort.

This commit establishes Python modules for discovering Firefox
release metadata and for persisting it in a SQLite database.

Release discovery works by scraping archive.mozilla.org. Essentially,
given a date range, we fire up a thread pool and spider that site,
looking for relevant files. Important release metadata is returned
as generic Python objects.

The release database is essentially a persistent store for the
data structures returned from release discovery. You can literally
feed the output of discovery to an import function to populate the
database. Duplicate insertions are handled properly.

A simple CLI script to triggering the release scraper and performing
basic actions has been added.

Currently, the scraper supports discovering Nightly from pure scraping
and Beta/Release from tags. The tags code isn't as well tested. And due
to the way the beta and release repos handle merges, walking the dag
to answer certain questions is difficult. So for all intents and
purposes we only care about the Nightly scraping code at this point.

This code should be considered alpha quality. I think it mostly works.
But the implementation of some parts (especially the database) is most
definitely a work in progress. I expect it to be iterated on heavily.

The good news is that scraping should produce a deterministic database.
So if we need to nuke the database in the future, you just scrape with
a fresh database and everything should "just work."

MozReview-Commit-ID: 7LRc1p9SDfy

--HG--
extra : rebase_source : b11a63764fc27e427fac5d54da00fc1aa54b8fa8