firefoxtree: add test for unknown revision (bug 1377081); r=glob
Currently, we don't filter entries in the firefoxtrees file against
repository contents. As a result, we may inject a reference to a
revision that doesn't exist or is filtered. This will cause a
run-time error as soon as we attempt to resolve the node referenced
by a namespace entry.

Add tests for this.

MozReview-Commit-ID: 2ec2ng5F7eC

--HG--
extra : amend_source : 6f37d6fc0c39dc36ce821ea2fe962fab5cb10a40