firefoxtree: ignore unknown revisions (bug 1377081); r=glob
As the previous tests demonstrated, we currently abort if a
reference to an unknown node leaks into the namespace. To
prevent this, we filter entries in the .hg/firefoxtrees
file at read time against known nodes.

There are side-effects to this change. For example, we'll silently
discard unknown nodes on next write. I think this is acceptable.
I'm trying to think of a scenario where silently discarding
references to old nodes is a bad thing. If a node is removed
(or hidden), we should drop our reference to it. Worst case,
if the node is re-introduced, the `hg pull` may re-populate the
firefoxtrees file if that node is still a head.

MozReview-Commit-ID: 3W294OAfhee

--HG--
extra : rebase_source : a060fb1f6c79b222329386fb5f93bbd1e05fab2b
extra : amend_source : fa49853edc00138d6dbe5dd5d249a7d0d302b6fd