hooks: establish extension for hooks and port single root hook; r=RyanVM
Currently, hooks are defined as individual Python modules. Hooks are
installed either in the global hgrc or in per-repository .hg/hgrc files.

This is sub-optimal for a few reasons.

First, we commonly have inconsistently installed hooks for repos of
the same class, like Firefox repos. We'll install a hook in one repo
but forget to install it in another. This can lead to obvious badness
and cause fire drills.

Second, because each hook is running independently, we lose some
performance. If a hook needs to do something like examine manifest
content, we could be resolving the same manifest fulltext in multiple
hooks. On repos the size of Firefox, this can amount to seconds of
CPU. This performance loss due to the existing hooks "architecture"
has been used to justify not enabling some "expensive" hooks.

In this commit, we establish a new "framework" for declaring and
running hooks. Hooks are known as "checks" and they are implemented
as a class. The API is designed in such a way that it encourages
checks to have optimal performance outcomes. e.g. by forcing checks
to examine each changeset in isolation, discrete checks can reuse
cached state (such as manifest data) from a previous check.

Instead of registering hooks as hooks, we create an extension.
The extension registers a hook internally which acts as a driver
for our registered classes. The use of an extension buys us a bit
of flexibility. We were already using Python-based hooks, which
are essentially extensions anyway, so this isn't a major departure.

One of the things we lose from explicit hooks is the ability to
add and remove hooks on a per-repo basis. To compensate for this,
we implement the mechanism via custom mozilla.check.* config options.

To prove all the code works, we port the single root hook to our new
framework.

MozReview-Commit-ID: C04Af2dct7w

--HG--
extra : rebase_source : 01a0e92d9ef93707e00635a82798732262b86026
extra : amend_source : da32f6efc5d995f7edd628bcbee122563991d145
extra : histedit_source : 0a558df745f8959b7a28c0d51fbd2d7232a88c4c