mozhg: add function to resolve repository metadata; r=glob
We often want to make decisions about what to do depending on the
"flavor" of a repo. Let's establish a function for resolving all
kinds of repository metadata to help with this.

As part of the introduction of this function, code in hooks has
been refactored to use the new API.

MozReview-Commit-ID: GU3oPAv1Y5v

--HG--
extra : rebase_source : 2971abda10a08f27b82824b48dac1c9a5dfc59fd