global: mass upgrade to Mercurial 4.2.3 (bug 1389216)
Previously, all Mercurial installs were updated to 4.2.2 or a custom
build of 4.2.2+ in preparation for the vulnerability fixes in 4.2.3.

This commit undoes all the hacks required for that and universally sets
the version to 4.2.3.

Mercurial screwed up their 4.2.3 tarball upload to PyPI and it can
never be uploaded there. So we upload a copy to our S3 bucket for random
packages and tell pip how to find Mercurial there. We can undo this
hack once a valid tar.gz is available on PyPI for a future release.

--HG--
extra : amend_source : e40515033051604d466297aaa063faeeffa0bd6c