robustcheckout: more robustly detect local store (bug 1391424); r=glob
As part of implementing more sanitization for the local store, I
discovered that we only examine the store repo if a working directory
is available. In many scenarios, we may have a store repo but no
working directory.

In this commit, we resolve the path to the store using the same
mechanism that Mercurial's pooled storage uses. This guarantees
that we'll find a local store even if we don't have a working
directory pointing to it.

As the locks test change shows, we now correctly detect a store
lock when there isn't a working directory.

This change does result in a redudant "lookup" wire protocol
command being issued. That's why we had to increase the request
counter in test-server-failure.t.

MozReview-Commit-ID: 3h17NesgxN3

--HG--
extra : rebase_source : 2c52bbf6fb9d67521eeeef27bc36e5c7c3ee8c89
extra : amend_source : db202125d8a8aeea6c4c637fc9acfe6bc01bcf6f