robustcheckout: validate requirements in store repo (bug 1391424); r=glob
For reasons I can't explain, we're seeing empty repos or repos without
.hg/requires files in automation. I suspect some process is attempting
to clone a repo, fails, and the .hg directory is somehow orphaned.

When Mercurial sees an existing .hg directory without a "requires"
file, it assumes it is an ancient repository format and it happily
starts writing new data to it.

One of the features not in this ancient repository format is the
ability to shorten paths in the store to avoid long path length limits
on common filesystems and operating systems. e.g. the 260 max path
length for various APIs on Windows.

The missing features also make some repository operations slower.

We should never encounter these ancient repository formats in the
wild. So, this commit teaches robustcheckout to verify that an
existing shared repo has modern requirements. If it doesn't, we
blow away that share and the working directory.

MozReview-Commit-ID: 4Me9vm2JMFy