testing: treat hooks as an extension directory (bug 1391366); r=glob
Previously, we treated the hooks directory as a one-off. With recent
changes, hooks are converting to an extension. So, this one-off is no
longer appropriate.

In this commit, we stop treating hooks as a one-off. We replace the
"hook" test type with "hg."

We emit the hooks directory as an extension directory. This means
we now discover tests as part of the extensions mechanism. And,
extension version compatibility is now detected. So we no longer
attempt to run hooks tests with versions of Mercurial not marked
as compatible by the extension.

MozReview-Commit-ID: 90TYDTj4z9K

--HG--
extra : amend_source : 22e7eef311ff74101e74e36453dde22b4a8a133e
extra : histedit_source : fc66180709d51afd82cd7298c6e3e8dd48ed6f81