hghooks: don't mark extension as compatible with 4.3 (bug 1391366); r=glob
Now that the hghooks directory is treated as an extension, this value
controls which versions of Mercurial the tests are run with. Since the
tests have various failures in 4.3, let's punt on fixing them to
another time by marking the extension only as compatible with 4.2.
This removes a large chunks of the remaining test failures in 4.3.

MozReview-Commit-ID: 6MjaG7X8CIA

--HG--
extra : histedit_source : 830496ca546a54c4beedc0fc2a50d6651f09444e