mozhg: don't use deprecated bookmarks APIs (bug 1391366); r=glob
Mercurial 4.3 issues a deprecation warning when not using the
applychanges() API. Switch the code to use the modern API when it is
available.

With this change, all non-Docker tests are no clean with 4.3 for
extensions marked as compatible with 4.3. There are still a number
of failures in extensions not marked as 4.3 compatible.

MozReview-Commit-ID: 3gVcHdnlxnf