testing: instantiate a docker.DockerClient
The previous commit moved everything to use self.api_client in
preparation for this change.

We now instantiate a high-level DockerClient and store it under
self.client. We don't yet use the new high-level API. But if we
have a need to, it is available.

--HG--
extra : amend_source : 3dbee5723f225f9bc49f392e935052e02859a537