reviewboard: mark extensions as compatible with 4.3 (bug 1413628); r=smacleod
All tests now pass on 4.3.

Strictly speaking, we're not sure if the server works with 4.3 because
we haven't upgraded the server to Mercurial 4.3. But we need to bump the
compatibility of both extensions so the test runner runs tests for 4.3.

MozReview-Commit-ID: I0k0x5qfTTO

--HG--
extra : rebase_source : 78aaaedc72ff8a63263f391557cec15f4465533c