reviewboard: avoid devel warning in Mercurial 4.3 (bug 1413628); r=smacleod
Mercurial 4.3+ will issue a devel warning if you pass a default
argument to look up the config value of an option with a static
default value.

There doesn't appear to be a good way to suppress the warning in
code aside from not passing the default argument. So we change
the code.

MozReview-Commit-ID: 9RbEMrq7IMZ

--HG--
extra : rebase_source : 86a2ecfc4cdfaf55956cf8ca03af764aa78e3f9c