reviewboard: register config options (bug 1413628); r=smacleod
Mercurial 4.4 softly requires config options be registered via a central
registrar. If they aren't, a devel warning is issued. The test runner
enables all devel warnings by default (which is a good idea). This
means we get new output and test failures for attempting to access a
config option that isn't registered.

When options are registered, Mercurial also issues a devel warning when
the caller accessing the option specifies a default value. This is
because the default value is defined as part of the option and doesn't
need to be specified at the call site.

Porting an extension to this mechanism is a bit annoying because old
Mercurial versions need to specify a default value (since the option
isn't registered) and new Mercurial versions would issue a warning
if a default argument is specified. The workaround is to set the
default value of "dynamicdefault." This special value means "the
default is dynamic" and ui.config* functions won't issue a devel
warning if a default value is specified.

When support for older versions of Mercurial are dropped, the proper
default value can be registered and the "default" argument can be
dropped from ui.config* callers.

This commit registers all config options used by the reviewboard
extension. By doing so, most test failures in 4.4 due to new
"devel-warn:" lines go away. This leaves a handful of legitimate
test failures.


MozReview-Commit-ID: KjS9MRuc2KD

--HG--
extra : rebase_source : 11defe73e5dd2824ea6d0d1d0a6a716745f5d3a7
extra : source : 8f281d062ec4e463cdf4374b1b7cdf0943d09749