testing: use a treestatus-legacy repo which vendors in dependencies (bug 1419998) r=gps
the old revision of treestatus we use in the dev/test environment references a
git submodule that no longer exists.

i've created a new repository at https://github.com/globau/treestatus-legacy
which contains the specific revision of treestatus that we use, as well as
vendoring in the libs from the submodule.

MozReview-Commit-ID: AMNty6mh41r

--HG--
extra : rebase_source : 5a438eaa66291855f26c20321090e8981ac50915