hgmo: Remove 'pile of poo' emoji from hg.m.o (Bug 1353195) r=gps
This patch removes the pile of poo emoji displayed on backed out changesets
in favour of the black skull and crossbones emoji. The background for the
warning is also changed to red to further highlight that the patch
has been backed out.

MozReview-Commit-ID: KAMbddrmEZt

--HG--
extra : rebase_source : 9364c344eb4580881427295ba554dd9dc8af24d0