testing: ping less and uglify-js in create-test-environment (bug 1410310)
If we're pinning them in Docker, we should be pinning them in the
test environment as well.

Not pinning them was resulting in errors running
`ln -s` because modern versions of uglify-js don't have the
bin/uglifyjs file and `ln -s` was complaining about linking to
a target that doesn't exist.