hgtemplates: port gitweb theme changes to gitweb_mozilla (bug 1423741); r=glob
This commit is a manual port of all gitweb theme changes from 4.2 to 4.3.

Most of the churn is around filelog and annotate views (the mechanism
for line-level annotate was refactored a bit in 4.3).

A CSS class for "footers" in some pages was also changed.

MozReview-Commit-ID: 5SGJvcg7jEl

--HG--
extra : rebase_source : 477de5c11ec8a8773ea386e859a8ee340d281aed