ansible: upgrade Mercurial system package to 4.3.3 (bug 1423741); r=glob
See the inline comment for an explanation for why we didn't touch
CentOS 6.

This will update the system Mercurial on hgssh, which we haven't yet
updated to 4.3. However, nothing of importance should be using the
system Mercurial. So it shouldn't matter.

MozReview-Commit-ID: GXKY4agDs4d

--HG--
extra : rebase_source : aaf55b262df88836469ec342a44d84e9604dda98