configwizard: remove pager configs on Mercurial 4.2+ (bug 1362547); r=glob
The pager extension was moved to core and enabled by default in
Mercurial 4.2. Furthermore, the presence of the legacy extension
and configs enables the old behavior of pager, which isn't desirable.

The default configuration for all intents and purposes should just
work. So, this commit changes the code to remove the extension
and pager.attend* config options on 4.2+. The only config options
that a user may wish to set in 4.2+ are pager.pager and pager.ignore.
Both of these are unchanged by the config wizard.

Because the enabling of `hg wip` installs a pager config, we had to
update that code and test as well.

MozReview-Commit-ID: GBNJB0mI2ZM

--HG--
extra : rebase_source : d5cf36c92a37ca6a322f4bec40f0ddd1e1be922e
extra : amend_source : 402d5119d45ee6d598445c409343e5bcf0670f25