firefoxreleases: use config registrar (bug 1424024)
Mercurial 4.3 introduced a registrar for config options. It allows
config options to be centrally defined, along with a default value.

Mercurial 4.4 issues a developer warning if you a) access an
unregistered config option b) specify a default value on access.

The latter property is a bit annoying because it breaks compatibility
with older Mercurial versions, which don't have a registrar and need
a default value argument. The workaround is to register the config
option as having a "dynamic" default value. We use that feature here.
However, we actually have a truly dynamic default value. So our
extension is "fully up to code."

With this change, the extension is now marked as compatible with
Mercurial 4.4.

--HG--
extra : amend_source : 714c4792664616fbaafc145c1cf13f3f646a9440
extra : histedit_source : b3f6251647bdf160ab639815a2ded791a7e941c5