scripts: script to upgrade Mercurial repos (bug 1351859)
This script essentially runs `hg debugupgraderepo` for multiple
repositories.

The intent of the script is to use it to bulk upgrade repos to
generaldelta. But it could be used for other upgrades in the future.

It supports parallel execution so it completes faster.

--HG--
extra : rebase_source : beed3867cf9460056aeff4ee95fa073d5dcad452