autoland: move setting db user's password from schema.sql to createdb.sh
Setting the db user's password to a hardcoded value in schema.sql is in the way
of using that file to initialise a database's schema.  Move it into the
`createdb.sh` script, which already creates the user and the database.

--HG--
extra : histedit_source : 89162bee43bf6b3396cc27d5fe7c7769d11f9133