Bug 1429261 - Do not fall back to using bookmark name, r=gps
The way the fallback was implemented resulted in original patches (as in, ones that had never yet been obsoleted) would default to the bookmark name, which is never what you want when using evolve, at least.

Rather than fixing it, this patch removes the bookmark fallback. Using a bookmark name as a persistent patch ID is very error-prone. It only works when you have a single patch under a bookmark, and otherwise is a giant footgun.