mozhg: adapt to Mercurial 4.3 API for querying config items (bug 1394898); r=glob
If we pass the default value for a built-in config item in
Mercurial 4.3+, we get a developer warning.

Change the code to not pass in a default value when the config
item is registered to avoid this warning. This avoids a whole
host of config warnings throughout the test suite.

Differential Revision: https://phabricator.services.mozilla.com/D448

--HG--
extra : rebase_source : 505f3024803b98650082223718ea43b27ae7720a
extra : amend_source : d2851de236e335a7d6ba49e36c38ed15dfe808e4