hghooks: Update dom peers in prevent_webidl_changes.py (Bug 1428785) r=gps
Some of the entries in the DOM_peers and DOM_authors lists have left
Mozilla, and some of the new DOM peers have yet to be added. This
commit adds the new peers and removes the former. Testing the hook
is now done with 'baku' as the test reviewer in place of 'jst'.

MozReview-Commit-ID: 3i25N3Zcti1

--HG--
extra : rebase_source : cb38286d69b2e6c6963749505b0f51a7f17ca3d2