hgserver: hook to prevent most repository changes on hgweb (bug 1439795); r=glob
Repositories on hgweb machines should be read-only except to the
replication mechanism.

To reinforce this desired property, this commit introduces a global
hook running as part of the pretxnchangegroup and precommit hooks
that attempts to make repositories read-only.

Local commits are banned outright.

Changegroups can only be applied if they are coming from `hg pull`
operations or are the result of an `hg strip`.

We still allow repository writes via the "pushkey" protocol. But this
is only because there isn't a good way for hooks to know what the
source of a "pushkey" request was. It also doesn't help that the
replication mechanism effectively calls `hg debugpushkey` to replicate
pushkey events. But something is better than nothing.

In addition to the added hooks tests, I verified that vcsreplicator
and Docker-based tests for hg.mo functionality continue to pass.

MozReview-Commit-ID: 8JBQDrC0sNd

--HG--
extra : amend_source : f663336f0b9a2747cf3d69691082cb0aa898d533