ansible/hg-web: install custom build of Mercurial 4.4 (bug 1440810); r=glob
We want to use a custom build of Mercurial 4.4 for hgweb to work
around an upstream bug.

The source distribution is private. (A copy is currently on the
moz-packages S3 bucket but the object is private.) We copy the
source distribution from the extra-files/ directory of v-c-t to
/var/tmp/pip. The pip requirements file is taught to look in
/var/tmp/pip for files.

This will make it a little annoying to test hgweb. If you need to
test hgweb on top of this change, you can revert this changeset
easily enough. Alternatively, you could copy the mercurial-4.4.2.tar.gz
file to extra-files/ and change references to the filename and
hash. But this hack should be short enough lived and I don't
think many people will get tripped up by it. So I don't think there's
a need to code a workaround.

MozReview-Commit-ID: 52841krY3SQ

--HG--
extra : rebase_source : 150ef299beb5b4bad4cc493364ef9a4824e51f92