serverlog: perform garbage collection on every request (bug 1443984); r=sheehan
hgwebdir is currently leaking repository objects. This can lead to OOM
on hgweb machines in production.

The leaks are likely due to a cycle in repository objects. Those
leaks likely won't get fixed until 4.6 at the earliest.

Since memory leaks are effectively a fact of life at this
juncture, let's mitigate their existence by forcing a garbage
collection at the end of every request.

A similar patch to do this is proposed upstream. Worst case, both us
and core perform a collect. The 2nd collect should be very fast.

We implement this in the serverlog extension because monkeypatching
hgweb is hard and this extension already does it.

MozReview-Commit-ID: 3HDarYrDF3J