vcttesting: Remove call to repo-push from create-repo docker script r=gps
The docker test suite contains a helper script "create-repo" which is
invoked to initialize a repo on the hgssh master, set the correct
permissions and replicate to the hgweb mirrors. Now that vcsreplicator
exists, we don't need to push the repo to the mirrors.

Differential Revision: https://phabricator.services.mozilla.com/D838

--HG--
extra : amend_source : 111f25b1f2b31158fff9940b6206cd16bebf1fe0