ansible/hg-web: use ansible_processor_vcpus r=fubar
ansible_processor_cores lists the number of cores per processor.
If you care about physical core count, you need to multiply this
number by ansible_processor_count. And if you want the number
of virtual CPUs, you need to multiply that by
ansible_processor_threads_per_core!

We were failing to do this and assuming we had 2 hyperthreaded CPUs.
This assumption was true for hgweb11-14. But hgweb15 and 16 only have
a single socket and have 12 virtual CPUs instead of 24. This meant
we were allocating 24 WSGI workers on 12 virtual CPUs. This caused CPU
load on hgweb15 and 16 to be ~2x of the other servers.

This commit switches the ansible logic to use
ansible_processor_vcpus for CPU counts.

wsgi_processes did not change for hgweb11-14. It was adjusted from 24
to 12 for hgweb15 and 16.

httpd_server_limit and httpd_max_clients also changed from 30 to 29,
due to rounding. This should be fine.

Differential Revision: https://phabricator.services.mozilla.com/D1044