reviewboard: mark extension as compatible with Mercurial 4.6 r=sheehan
And drop support for 4.2 given our policy to only support the last 4
releases.

All tests pass with 4.6rc1, so this should be safe.

The server-side bits likely don't work with 4.6. But the server bits
should only need to work on 4.4, which is what we're running on
reviewboard-hg. So we can defer upgrading that code until later
(potentially never if reviewboard-hg just rots until it is turned off).

Differential Revision: https://phabricator.services.mozilla.com/D1099