vcsreplicator: revert changes from 22cebb54cab1
The path reference to bootstrap.py was busted. I'm tempted to update it
to use the proper relative path. However, I believe this code runs in a
virtualenv and may not be executed from v-c-t. So relative paths may not
work.

Let's just revert to the old try..except methods. Except we add a
getattr() to force the ImportError.