hghooks: measure and log time spent in individual hooks r=glob
Mercurial measures the time that individual hooks take. But our hooks
"framework" appears to Mercurial as a single hook. We thus lose
visibility into how long each of our hooks take.

This commit implements support for measuring time spent in the
sub-hooks that we run. This should shed visibility into which of our
hooks are taking long to execute.

Differential Revision: https://phabricator.services.mozilla.com/D1074