hghooks: short circuit subrepos check if files missing (bug 1457342) r=glob
Previously, we would read the manifest for each context and look for
the presence of an .hgsub or .hgsubstate file. Reading manifests can
be expensive for large manifests (such as those found in Firefox
repos).

I suspect that reading manifests is making large pushes to Firefox
repos slower than they could be.

This commit implements an optimization to make subrepo checking
faster. Essentially, when the hook starts, it looks to see if there
are *any* revisions for a .hgsub or .hgsubstate file. If not, then
no revision contains this file and we can avoid a potentially
expensive manifest resolve and lookup.

Differential Revision: https://phabricator.services.mozilla.com/D1073