vcsreplicator: upgrade kafka-python to 1.4.2 (Bug 1453089) r=gps
Upgrade kafka-python from v0.9 to v1.4.2. This will allow leveraging
new APIs so vcsreplicator isn't built off deprecated code. A few
of the names have been moved in the new version of the library, so
we rename things accordingly.

kafka-python 1.4.2 doesn't seem to like Mercurial's demand import
scheme, so we deactivate it in the vcsreplicator extension.

This commit includes some test fixes, and appends
` | grep ^vcsreplicator.consumer` to a number of the test cases.
The tests are breaking due to "UnknownTopicOrPartitionError" stack
traces in the consumer.log output, however the log output following
the error indicates the topic and partition have been found and all
is well. It is most likely that a race condition has been introduced
while upgrading to Kafka 1.1. So we filter the output for only the
consumer log messages we expect. If a topic or partition was truly
unknown to the consumer, the test will still fail as the log message
we anticipate in the test will not appear 8 times (once for each
partition). It is also worth noting that there are still test failures
despite this change, however the amount of noise in the output is
significantly reduced.

Differential Revision: https://phabricator.services.mozilla.com/D927