vcsreplicator: move kafka message deserialization into new function r=gps
In the new KafkaConsumer API, messages are deserialized via a deserialization
function. The vcsreplicator.Consumer class has a few lines of code
that turn the message value into a Python dict. This is essentially
the equivalent of the deserialization function, so move this functionality
out of the instance method get_message and have get_message call this
instead.

Differential Revision: https://phabricator.services.mozilla.com/D1154