configwizard: drop marked compatibility with Mercurial <4.3; r=sheehan
I can't remember why we were keeping around marked support for older
Mercurial versions. This extension is special in that we want it to work
with older Mercurial versions. But I can't recall why we were keeping
the old testedwith values. If we want to test with old versions, we can
always run tests with `--with-hg /path/to/old/version` and it should
"just work."

Differential Revision: https://phabricator.services.mozilla.com/D1206