docs: generate requirements file with pip-compile; r=sheehan
pip-compile is a useful command from pip-tools. It allows you to define
a source file containing packages and it automagically determines
dependencies and obtains package hashes. It is super useful for managing
pip requirements files because you no longer need to manually update
hashes or track which packages are "primary" versus dependencies of
another.

In this commit, we introduce a source spec file for
docs-requirements.txt. We include all dependency packages and their
versions for now so the [now] auto-generated docs-requirements.txt
has less churn in this commit.

Differential Revision: https://phabricator.services.mozilla.com/D1163