hghooks: use full nodes in more places; r=sheehan
Mercurial 4.6. is a bit more strict about what values can be passed to
repo[x] (which calls changectx.__init__ in many cases). Since repo[x] is
an internal API, 4.6 issues a deprecation warning if the passed node
is not a full node.

The change changed in this commit was previously passing short hex
nodes into repo[x], thus triggering deprecation warnings on Mercurial
4.6.

This commit refactors the code to use full binary nodes internally
and to only convert to short hex nodes for presentation purposes.

Differential Revision: https://phabricator.services.mozilla.com/D1223