vcsreplicator: add hgssh bootstrap procedure r=gps
This commit adds the hgssh component of the vcsreplicator hgweb
bootstrapping procedure. First we gather the most recent offsets for
each partition on the topic. Then we use hgmolib/find_hg_repos to
get a list of every repository on hg.m.o. For each repo in this list,
we publish a `replicatesync` message. Then we gather the most recent
offsets after all messages have been published, which gives us a range
of offsets for which the replication process is considered valid. Once
this is completed, a JSON object is printed to stdout for collection
via Ansible. This JSON object will be passed to the new hgweb head for
the second piece of the bootstrap procedure.

To test this procedure, we turn off vcsreplicator on hgweb0 and create
some repositories for replication. We then run the bootstrap procedure
and verify the state of the Kafka topic, as well as the format of the
emitted JSON. The test will be extended to test for successful
replication in the next commit of this series.

Differential Revision: https://phabricator.services.mozilla.com/D1155