vcsreplicator: remove support for creating repos without generaldelta; r=sheehan
Repositories are created with generaldelta by default as of some
Mercurial version older than the oldest Mercurial we are running in
production. And, every repository on hg.mozilla.org has been converted
to generaldelta.

So, we no longer need special code for controlling whether repositories
are created with or without generaldelta.

vcsreplicator's (failint) test-init.t was updated accordingly.

The change to test-clonebundles.t does drop some test coverage. But it's
for a branch we should never encounter in production since all repos are
generaldelta. I don't think it's worth keeping the code around to
facilitate testing it.

Differential Revision: https://phabricator.services.mozilla.com/D1320